requirements.txt
aiomysql==0.0.21
aniso8601==9.0.1
bcrypt==3.2.0
cffi==1.14.5
click==8.0.1
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-RESTful==0.3.9
importlib-metadata==4.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.0
pandas==1.2.5
plum-dispatch==1.1.1
pycparser==2.20
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
typing-extensions==3.10.0.0
uWSGI==2.0.19.1
Werkzeug==2.0.1
zipp==3.4.1
